ci: disable cleanup on ci

+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/helpers/extensions/MetadataSetupExtension.java
@@ -137,13 +137,14 @@ public class MetadataSetupExtension
 
     @Override
     public void close()
-        throws Throwable
     {
-        TestCleanUp testCleanUp = new TestCleanUp();
+        if ( TestConfiguration.get().shouldCleanUp() ) {
+            TestCleanUp testCleanUp = new TestCleanUp();
 
-        iterateCreatedData( id -> {
-            testCleanUp.deleteEntity( createdData.get( id ), id );
-        } );
+            iterateCreatedData( id -> {
+                testCleanUp.deleteEntity( createdData.get( id ), id );
+            } );
+        }
 
     }
 }
